backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.3
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0
